requirements.txt
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
outcome==1.2.0
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.0
pytube==15.0.0
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
spotipy==2.23.0
trio==0.22.2
trio-websocket==0.10.3
uritemplate==4.1.1
urllib3==1.26.16
wsproto==1.2.0